# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 64230e6c87ddb2742c6dd162
  title: schoolmng
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/exam/{id}":
    delete:
      x-skaffolder-id: 642311f587ddb2742c6dd23c
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 642310e387ddb2742c6dd206
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 642311f587ddb2742c6dd239
      x-skaffolder-name: get
      x-skaffolder-id-resource: 642310e387ddb2742c6dd206
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/exam"
    post:
      x-skaffolder-id: 642311f587ddb2742c6dd237
      x-skaffolder-name: update
      x-skaffolder-id-resource: 642310e387ddb2742c6dd206
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/exam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/exam"
  "/exam/findBy_course/{key}":
    get:
      x-skaffolder-id: 642311e887ddb2742c6dd22c
      x-skaffolder-name: findBy_course
      x-skaffolder-id-resource: 642310e387ddb2742c6dd206
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_course
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_course
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_course/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_course
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/findBy_student/{key}":
    get:
      x-skaffolder-id: 642311cf87ddb2742c6dd213
      x-skaffolder-name: findBy_student
      x-skaffolder-id-resource: 642310e387ddb2742c6dd206
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_student
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_student
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_student/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_student
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/findBy_teacher/{key}":
    get:
      x-skaffolder-id: 642311dc87ddb2742c6dd220
      x-skaffolder-name: findBy_teacher
      x-skaffolder-id-resource: 642310e387ddb2742c6dd206
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: findBy_teacher
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_teacher
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_teacher/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_teacher
      tags:
        - exam
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/exam"
  "/exam/{id}/validate":
    post:
      x-skaffolder-id: 642313f087ddb2742c6dd245
      x-skaffolder-name: validate
      x-skaffolder-id-resource: 642310e387ddb2742c6dd206
      x-skaffolder-resource: exam
      x-skaffolder-crudAction: 
      x-skaffolder-crudType: 
      x-skaffolder-description: This API is used to validate the exam
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: Boolean
      x-skaffolder-url: /{id}/validate
      x-skaffolder-roles:
      summary: This API is used to validate the exam
      tags:
        - exam
      parameters:
        - name: id
          x-skaffolder-type: String
          in: path
          description: id of the exam
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    exam:
      x-skaffolder-id: 642310e387ddb2742c6dd206
      x-skaffolder-id-db: 64230e6c87ddb2742c6dd163
      x-skaffolder-id-entity: 642310e387ddb2742c6dd203
      x-skaffolder-url: /exam
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        place:
          type: string
          x-skaffolder-id-attr: 642310e387ddb2742c6dd205
          x-skaffolder-type: String
        score:
          type: number
          x-skaffolder-id-attr: 642310e387ddb2742c6dd204
          x-skaffolder-type: Number
        valid:
          type: boolean
          x-skaffolder-id-attr: 642314bc87ddb2742c6dd247
          x-skaffolder-type: Boolean
      x-skaffolder-relations:
        _course:
          x-skaffolder-id: 6423112387ddb2742c6dd209
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 642310e387ddb2742c6dd203
          x-skaffolder-ent2: 642310cb87ddb2742c6dd1fe
        _student:
          x-skaffolder-id: 6423112387ddb2742c6dd208
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 642310e387ddb2742c6dd203
          x-skaffolder-ent2: 6423108787ddb2742c6dd1f0
        _teacher:
          x-skaffolder-id: 6423112387ddb2742c6dd20a
          x-skaffolder-type: 1:m
          x-skaffolder-ent1: 642310e387ddb2742c6dd203
          x-skaffolder-ent2: 642310b287ddb2742c6dd1f8
      required:
        - _id
...